Split shared JUnit parser from the backend adapter; name adapters by suite
The frontend (jest-junit) is also JUnit, so a file named adapters/junit.ts is
ambiguous — it's really the backend adapter. Separate the two concerns:

- junit.ts: the pure, shared JUnit XML parser (parseJunit), reused by the
  backend and the future frontend adapter; its tests move here with it.
- adapters/backend.ts: normalizeBackendJunit — hawk's *_test.xml discovery on
  top of the shared parser. Frontend later adds adapters/frontend.ts reusing
  junit.ts; cypress (not JUnit) stays separate.
- Rename the entrypoint report-junit.ts -> report-backend.ts (be and fe will
  have separate entrypoints), updating the composite action and root script.
