+++ package.json
@@ -514,7 +514,7 @@
     "embedding-sdk:clean:tsbuildinfo": "rm -rf resources/embedding-sdk/.tsbuildinfo",
     "eslint-fix": "bun run lint-eslint --fix",
     "find-never-defined-css-variables": "tsx bin/find-never-defined-css-variables/find-never-defined-css-variables.ts",
-    "report-junit-failures": "bun release/ci-conductor/src/report-junit.ts",
+    "report-junit-failures": "bun release/ci-conductor/src/report-backend.ts",
     "generate-cypress-html-report": "mochawesome-merge cypress/reports/mochareports/*.json > cypress/reports/cypress-test-report.json && marge cypress/reports/cypress-test-report.json -o cypress/reports --inline",
     "generate-openapi": "clojure -M:run:ee generate-openapi-spec",
     "lint-openapi": "redocly lint resources/openapi/openapi.json --skip-rule security-defined --skip-rule operation-operationId --skip-rule no-unused-components --skip-rule info-license --skip-rule no-path-trailing-slash --skip-rule no-identical-paths --skip-rule no-ambiguous-paths --skip-rule path-parameters-defined --skip-rule spec-components-invalid-map-name",
